WaykNow-ps: fix cases where directory and files do not exist prior to running the script

+++ WaykNow/Public/WaykNowLicense.ps1
@@ -8,9 +8,16 @@ function Set-WaykNowLicense
 
     $licensePattern = '[A-Z0-9]{5}-[A-Z0-9]{5}-[A-Z0-9]{5}-[A-Z0-9]{5}-[A-Z0-9]{5}'
     $WaykNowInfo = Get-WaykNowInfo
+    $ConfigFile = $WaykNowInfo.ConfigFile
 
     if ($License -CMatch $licensePattern) {
-        $json = Get-Content -Path $WaykNowInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+
+        if (Test-Path $ConfigFile) {
+            $json = Get-Content -Path $ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+        } else {
+            $json = '{}' | ConvertFrom-Json
+        }
+
         if ($json.RegistrationSerial)
         {
             $json.RegistrationSerial = $License;
@@ -26,9 +33,11 @@ function Set-WaykNowLicense
             $json | Add-Member -Type NoteProperty -Name 'RegistrationSerial' -Value $License -Force
         }
 
+        New-Item -Path $(Split-Path $ConfigFile -Parent) -ItemType 'Directory' -Force
+
         $fileValue = $json | ConvertTo-Json
         $Utf8NoBomEncoding = New-Object System.Text.UTF8Encoding $False
-        [System.IO.File]::WriteAllLines($WaykNowInfo.ConfigFile, $fileValue, $Utf8NoBomEncoding)
+        [System.IO.File]::WriteAllLines($ConfigFile, $fileValue, $Utf8NoBomEncoding)
     } else {
         Write-Error "Invalid License Format"
     }
@@ -40,8 +49,13 @@ function Get-WaykNowLicense
     param()
 
     [WaykNowInfo]$WaykInfo = Get-WaykNowInfo
-    $json = Get-Content -Path $WaykInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
-    return $json.RegistrationSerial
+
+    if (Test-Path $WaykInfo.ConfigFile) {
+        $json = Get-Content -Path $WaykInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+        return $json.RegistrationSerial;
+    } else {
+        return $null
+    }
 }
 
 function Reset-WaykNowLicense
@@ -51,13 +65,15 @@ function Reset-WaykNowLicense
 
     [WaykNowInfo]$WaykInfo = Get-WaykNowInfo
 
-    $json = Get-Content -Path $WaykInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+    if (Test-Path $WaykInfo.ConfigFile) {
+        $json = Get-Content -Path $WaykInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
 
-    if ($json.RegistrationSerial) {
-        $json.RegistrationSerial = ''
-        $fileValue = $json | ConvertTo-Json
-        $Utf8NoBomEncoding = New-Object System.Text.UTF8Encoding $False
-        [System.IO.File]::WriteAllLines($WaykInfo.ConfigFile, $fileValue, $Utf8NoBomEncoding)
+        if ($json.RegistrationSerial) {
+            $json.RegistrationSerial = ''
+            $fileValue = $json | ConvertTo-Json
+            $Utf8NoBomEncoding = New-Object System.Text.UTF8Encoding $False
+            [System.IO.File]::WriteAllLines($WaykInfo.ConfigFile, $fileValue, $Utf8NoBomEncoding)
+        }
     }
 }
